services/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    AWS_ACCESS_KEY = os.getenv('AWS_ACCESS_KEY')
    AWS_SECRET_KEY = os.getenv('AWS_SECRET_KEY')
    AWS_BUCKET_NAME = os.getenv('AWS_BUCKET_NAME', 'background-removal-bucket')
    REGION_NAME = os.getenv('AWS_REGION', 'us-east-1')

settings = Settings()