requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
boto3==1.35.78
botocore==1.35.78
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.6
h11==0.14.0
idna==3.10
imageio==2.36.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
pydantic==2.10.3
pydantic_core==2.27.1
PyMatting==1.1.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
referencing==0.35.1
rembg==2.0.60
requests==2.32.3
rpds-py==0.22.3
s3transfer==0.10.4
scikit-image==0.24.0
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tifffile==2024.9.20
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
